Split Cardano.Api.LedgerState.Fold module
